remove old promql editor (#9452)
* remove old promql editor


* rename CMExpression by Expression
