+++ web/ui/react-app/src/pages/graph/ExpressionInput.tsx
@@ -1,15 +1,30 @@
-import React, { Component } from 'react';
-import { Button, Input, InputGroup, InputGroupAddon, InputGroupText } from 'reactstrap';
+import React, { FC, useState, useEffect, useRef } from 'react';
+import { Button, InputGroup, InputGroupAddon, InputGroupText } from 'reactstrap';
 
-import Downshift, { ControllerStateAndHelpers } from 'downshift';
-import sanitizeHTML from 'sanitize-html';
+import { EditorView, highlightSpecialChars, keymap, ViewUpdate, placeholder } from '@codemirror/view';
+import { EditorState, Prec, Compartment } from '@codemirror/state';
+import { indentOnInput, syntaxTree } from '@codemirror/language';
+import { history, historyKeymap } from '@codemirror/history';
+import { defaultKeymap, insertNewlineAndIndent } from '@codemirror/commands';
+import { bracketMatching } from '@codemirror/matchbrackets';
+import { closeBrackets, closeBracketsKeymap } from '@codemirror/closebrackets';
+import { highlightSelectionMatches } from '@codemirror/search';
+import { commentKeymap } from '@codemirror/comment';
+import { lintKeymap } from '@codemirror/lint';
+import { autocompletion, completionKeymap, CompletionContext, CompletionResult } from '@codemirror/autocomplete';
+import { baseTheme, lightTheme, darkTheme, promqlHighlighter } from './CMTheme';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faGlobeEurope, faSearch, faSpinner } from '@fortawesome/free-solid-svg-icons';
+import { faSearch, faSpinner, faGlobeEurope } from '@fortawesome/free-solid-svg-icons';
 import MetricsExplorer from './MetricsExplorer';
-import { Fuzzy, FuzzyResult } from '@nexucis/fuzzy';
+import { usePathPrefix } from '../../contexts/PathPrefixContext';
+import { useTheme } from '../../contexts/ThemeContext';
+import { CompleteStrategy, PromQLExtension } from 'codemirror-promql';
+import { newCompleteStrategy } from 'codemirror-promql/dist/cjs/complete';
 
-interface ExpressionInputProps {
+const promqlExtension = new PromQLExtension();
+
+interface CMExpressionInputProps {
   value: string;
   onExpressionChange: (expr: string) => void;
   queryHistory: string[];
@@ -17,235 +32,218 @@ interface ExpressionInputProps {
   executeQuery: () => void;
   loading: boolean;
   enableAutocomplete: boolean;
+  enableHighlighting: boolean;
+  enableLinter: boolean;
 }
 
-interface ExpressionInputState {
-  height: number | string;
-  showMetricsExplorer: boolean;
+const dynamicConfigCompartment = new Compartment();
+
+// Autocompletion strategy that wraps the main one and enriches
+// it with past query items.
+export class HistoryCompleteStrategy implements CompleteStrategy {
+  private complete: CompleteStrategy;
+  private queryHistory: string[];
+  constructor(complete: CompleteStrategy, queryHistory: string[]) {
+    this.complete = complete;
+    this.queryHistory = queryHistory;
+  }
+
+  promQL(context: CompletionContext): Promise<CompletionResult | null> | CompletionResult | null {
+    return Promise.resolve(this.complete.promQL(context)).then((res) => {
+      const { state, pos } = context;
+      const tree = syntaxTree(state).resolve(pos, -1);
+      const start = res != null ? res.from : tree.from;
+
+      if (start !== 0) {
+        return res;
+      }
+
+      const historyItems: CompletionResult = {
+        from: start,
+        to: pos,
+        options: this.queryHistory.map((q) => ({
+          label: q.length < 80 ? q : q.slice(0, 76).concat('...'),
+          detail: 'past query',
+          apply: q,
+          info: q.length < 80 ? undefined : q,
+        })),
+        span: /^[a-zA-Z0-9_:]+$/,
+      };
+
+      if (res !== null) {
+        historyItems.options = historyItems.options.concat(res.options);
+      }
+      return historyItems;
+    });
+  }
 }
 
-const fuz = new Fuzzy({ pre: '<strong>', post: '</strong>', shouldSort: true });
+const ExpressionInput: FC<CMExpressionInputProps> = ({
+  value,
+  onExpressionChange,
+  queryHistory,
+  metricNames,
+  executeQuery,
+  loading,
+  enableAutocomplete,
+  enableHighlighting,
+  enableLinter,
+}) => {
+  const containerRef = useRef<HTMLDivElement>(null);
+  const viewRef = useRef<EditorView | null>(null);
+  const [showMetricsExplorer, setShowMetricsExplorer] = useState<boolean>(false);
+  const pathPrefix = usePathPrefix();
+  const { theme } = useTheme();
 
-class ExpressionInput extends Component<ExpressionInputProps, ExpressionInputState> {
-  private exprInputRef = React.createRef<HTMLInputElement>();
+  // (Re)initialize editor based on settings / setting changes.
+  useEffect(() => {
+    // Build the dynamic part of the config.
+    promqlExtension
+      .activateCompletion(enableAutocomplete)
+      .activateLinter(enableLinter)
+      .setComplete({
+        completeStrategy: new HistoryCompleteStrategy(
+          newCompleteStrategy({
+            remote: { url: pathPrefix, cache: { initialMetricList: metricNames } },
+          }),
+          queryHistory
+        ),
+      });
+    const dynamicConfig = [
+      enableHighlighting ? promqlHighlighter : [],
+      promqlExtension.asExtension(),
+      theme === 'dark' ? darkTheme : lightTheme,
+    ];
 
-  constructor(props: ExpressionInputProps) {
-    super(props);
-    this.state = {
-      height: 'auto',
-      showMetricsExplorer: false,
-    };
-  }
+    // Create or reconfigure the editor.
+    const view = viewRef.current;
+    if (view === null) {
+      // If the editor does not exist yet, create it.
+      if (!containerRef.current) {
+        throw new Error('expected CodeMirror container element to exist');
+      }
 
-  componentDidMount(): void {
-    this.setHeight();
-  }
+      const startState = EditorState.create({
+        doc: value,
+        extensions: [
+          baseTheme,
+          highlightSpecialChars(),
+          history(),
+          EditorState.allowMultipleSelections.of(true),
+          indentOnInput(),
+          bracketMatching(),
+          closeBrackets(),
+          autocompletion(),
+          highlightSelectionMatches(),
+          EditorView.lineWrapping,
+          keymap.of([
+            ...closeBracketsKeymap,
+            ...defaultKeymap,
+            ...historyKeymap,
+            ...commentKeymap,
+            ...completionKeymap,
+            ...lintKeymap,
+          ]),
+          placeholder('Expression (press Shift+Enter for newlines)'),
+          dynamicConfigCompartment.of(dynamicConfig),
+          // This keymap is added without precedence so that closing the autocomplete dropdown
+          // via Escape works without blurring the editor.
+          keymap.of([
+            {
+              key: 'Escape',
+              run: (v: EditorView): boolean => {
+                v.contentDOM.blur();
+                return false;
+              },
+            },
+          ]),
+          Prec.override(
+            keymap.of([
+              {
+                key: 'Enter',
+                run: (v: EditorView): boolean => {
+                  executeQuery();
+                  return true;
+                },
+              },
+              {
+                key: 'Shift-Enter',
+                run: insertNewlineAndIndent,
+              },
+            ])
+          ),
+          EditorView.updateListener.of((update: ViewUpdate): void => {
+            onExpressionChange(update.state.doc.toString());
+          }),
+        ],
+      });
 
-  setHeight = (): void => {
-    if (this.exprInputRef.current) {
-      const { offsetHeight, clientHeight, scrollHeight } = this.exprInputRef.current;
-      const offset = offsetHeight - clientHeight; // Needed in order for the height to be more accurate.
-      this.setState({ height: scrollHeight + offset });
-    }
-  };
+      const view = new EditorView({
+        state: startState,
+        parent: containerRef.current,
+      });
 
-  handleInput = (): void => {
-    if (this.exprInputRef.current) {
-      this.setValue(this.exprInputRef.current.value);
-    }
-  };
+      viewRef.current = view;
 
-  setValue = (value: string): void => {
-    const { onExpressionChange } = this.props;
-    onExpressionChange(value);
-    this.setState({ height: 'auto' }, this.setHeight);
-  };
-
-  componentDidUpdate(prevProps: ExpressionInputProps): void {
-    const { value } = this.props;
-    if (value !== prevProps.value) {
-      this.setValue(value);
-    }
-  }
-
-  handleKeyPress = (event: React.KeyboardEvent<HTMLInputElement>): void => {
-    const { executeQuery } = this.props;
-    if (event.key === 'Enter' && !event.shiftKey) {
-      executeQuery();
-      event.preventDefault();
-    }
-  };
-
-  getSearchMatches = (input: string, expressions: string[]): FuzzyResult[] => {
-    return fuz.filter(input.replace(/ /g, ''), expressions);
-  };
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  createAutocompleteSection = (downshift: ControllerStateAndHelpers<any>): JSX.Element | null => {
-    const { inputValue = '', closeMenu, highlightedIndex } = downshift;
-    const autocompleteSections = {
-      'Query History': this.props.queryHistory,
-      'Metric Names': this.props.metricNames,
-    };
-    let index = 0;
-    const sections =
-      inputValue?.length && this.props.enableAutocomplete
-        ? Object.entries(autocompleteSections).reduce((acc, [title, items]) => {
-            const matches = this.getSearchMatches(inputValue, items);
-            return !matches.length
-              ? acc
-              : [
-                  ...acc,
-                  <ul className="autosuggest-dropdown-list" key={title}>
-                    <li className="autosuggest-dropdown-header">{title}</li>
-                    {matches
-                      .slice(0, 100) // Limit DOM rendering to 100 results, as DOM rendering is slow.
-                      .map((result: FuzzyResult) => {
-                        const itemProps = downshift.getItemProps({
-                          key: result.original,
-                          index,
-                          item: result.original,
-                          style: {
-                            backgroundColor: highlightedIndex === index++ ? 'lightgray' : 'white',
-                          },
-                        });
-                        return (
-                          <li
-                            key={title}
-                            {...itemProps}
-                            dangerouslySetInnerHTML={{ __html: sanitizeHTML(result.rendered, { allowedTags: ['strong'] }) }}
-                          />
-                        );
-                      })}
-                  </ul>,
-                ];
-          }, [] as JSX.Element[])
-        : [];
-
-    if (!sections.length) {
-      // This is ugly but is needed in order to sync state updates.
-      // This way we force downshift to wait React render call to complete before closeMenu to be triggered.
-      setTimeout(closeMenu);
-      return null;
-    }
-
-    return (
-      <div {...downshift.getMenuProps()} className="autosuggest-dropdown">
-        {sections}
-      </div>
-    );
-  };
-
-  openMetricsExplorer = (): void => {
-    this.setState({
-      showMetricsExplorer: true,
-    });
-  };
-
-  updateShowMetricsExplorer = (show: boolean): void => {
-    this.setState({
-      showMetricsExplorer: show,
-    });
-  };
-
-  insertAtCursor = (value: string): void => {
-    if (!this.exprInputRef.current) return;
-
-    const startPosition = this.exprInputRef.current.selectionStart;
-    const endPosition = this.exprInputRef.current.selectionEnd;
-
-    const previousValue = this.exprInputRef.current.value;
-    let newValue: string;
-    if (startPosition && endPosition) {
-      newValue =
-        previousValue.substring(0, startPosition) + value + previousValue.substring(endPosition, previousValue.length);
+      view.focus();
     } else {
-      newValue = previousValue + value;
+      // The editor already exists, just reconfigure the dynamically configured parts.
+      view.dispatch(
+        view.state.update({
+          effects: dynamicConfigCompartment.reconfigure(dynamicConfig),
+        })
+      );
     }
+    // "value" is only used in the initial render, so we don't want to
+    // re-run this effect every time that "value" changes.
+    //
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [enableAutocomplete, enableHighlighting, enableLinter, executeQuery, onExpressionChange, queryHistory, theme]);
 
-    this.setValue(newValue);
+  const insertAtCursor = (value: string) => {
+    const view = viewRef.current;
+    if (view === null) {
+      return;
+    }
+    const { from, to } = view.state.selection.ranges[0];
+    view.dispatch(
+      view.state.update({
+        changes: { from, to, insert: value },
+      })
+    );
   };
 
-  render(): JSX.Element {
-    const { executeQuery, value } = this.props;
-    const { height } = this.state;
-    return (
-      <>
-        <Downshift onSelect={this.setValue}>
-          {(downshift) => (
-            <div>
-              <InputGroup className="expression-input">
-                <InputGroupAddon addonType="prepend">
-                  <InputGroupText>
-                    {this.props.loading ? <FontAwesomeIcon icon={faSpinner} spin /> : <FontAwesomeIcon icon={faSearch} />}
-                  </InputGroupText>
-                </InputGroupAddon>
-                <Input
-                  onInput={this.handleInput}
-                  style={{ height }}
-                  autoFocus
-                  type="textarea"
-                  rows="1"
-                  onKeyPress={this.handleKeyPress}
-                  placeholder="Expression (press Shift+Enter for newlines)"
-                  innerRef={this.exprInputRef}
-                  {...downshift.getInputProps({
-                    onKeyDown: (event: React.KeyboardEvent): void => {
-                      switch (event.key) {
-                        case 'Home':
-                        case 'End':
-                          // We want to be able to jump to the beginning/end of the input field.
-                          // By default, Downshift otherwise jumps to the first/last suggestion item instead.
-                          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                          (event.nativeEvent as any).preventDownshiftDefault = true;
-                          break;
-                        case 'ArrowUp':
-                        case 'ArrowDown':
-                          if (!downshift.isOpen) {
-                            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                            (event.nativeEvent as any).preventDownshiftDefault = true;
-                          }
-                          break;
-                        case 'Enter':
-                          downshift.closeMenu();
-                          break;
-                        case 'Escape':
-                          if (!downshift.isOpen && this.exprInputRef.current) {
-                            this.exprInputRef.current.blur();
-                          }
-                          break;
-                        default:
-                      }
-                    },
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                  } as any)}
-                  value={value}
-                />
-                <InputGroupAddon addonType="append">
-                  <Button className="metrics-explorer-btn" title="Open metrics explorer" onClick={this.openMetricsExplorer}>
-                    <FontAwesomeIcon icon={faGlobeEurope} />
-                  </Button>
-                </InputGroupAddon>
-                <InputGroupAddon addonType="append">
-                  <Button className="execute-btn" color="primary" onClick={executeQuery}>
-                    Execute
-                  </Button>
-                </InputGroupAddon>
-              </InputGroup>
-              {downshift.isOpen && this.createAutocompleteSection(downshift)}
-            </div>
-          )}
-        </Downshift>
+  return (
+    <>
+      <InputGroup className="expression-input">
+        <InputGroupAddon addonType="prepend">
+          <InputGroupText>
+            {loading ? <FontAwesomeIcon icon={faSpinner} spin /> : <FontAwesomeIcon icon={faSearch} />}
+          </InputGroupText>
+        </InputGroupAddon>
+        <div ref={containerRef} className="cm-expression-input" />
+        <InputGroupAddon addonType="append">
+          <Button
+            className="metrics-explorer-btn"
+            title="Open metrics explorer"
+            onClick={() => setShowMetricsExplorer(true)}
+          >
+            <FontAwesomeIcon icon={faGlobeEurope} />
+          </Button>
+          <Button className="execute-btn" color="primary" onClick={executeQuery}>
+            Execute
+          </Button>
+        </InputGroupAddon>
+      </InputGroup>
 
-        <MetricsExplorer
-          show={this.state.showMetricsExplorer}
-          updateShow={this.updateShowMetricsExplorer}
-          metrics={this.props.metricNames}
-          insertAtCursor={this.insertAtCursor}
-        />
-      </>
-    );
-  }
-}
+      <MetricsExplorer
+        show={showMetricsExplorer}
+        updateShow={setShowMetricsExplorer}
+        metrics={metricNames}
+        insertAtCursor={insertAtCursor}
+      />
+    </>
+  );
+};
 
 export default ExpressionInput;
