+++ web/ui/react-app/src/pages/graph/Panel.test.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { mount, shallow } from 'enzyme';
 import Panel, { PanelOptions, PanelType } from './Panel';
-import ExpressionInput from './ExpressionInput';
 import GraphControls from './GraphControls';
 import { NavLink, TabPane } from 'reactstrap';
 import TimeInput from './TimeInput';
@@ -37,17 +36,6 @@ const defaultProps = {
 
 describe('Panel', () => {
   const panel = shallow(<Panel {...defaultProps} />);
-
-  it('renders an ExpressionInput', () => {
-    const input = panel.find(ExpressionInput);
-    expect(input.prop('value')).toEqual('prometheus_engine');
-    expect(input.prop('metricNames')).toEqual([
-      'prometheus_engine_queries',
-      'prometheus_engine_queries_concurrent_max',
-      'prometheus_engine_query_duration_seconds',
-    ]);
-    expect(input.prop('queryHistory')).toEqual([]);
-  });
 
   it('renders NavLinks', () => {
     const results: PanelOptions[] = [];
